measurement api finished

+++ src/validation/tailor-validation.js
@@ -5,10 +5,11 @@ const registerTailorValidation = Joi.object({
     store_name: Joi.string().max(100).required(),
     address: Joi.string().max(255).optional(),
     description: Joi.string().max(255).optional(),
-    longtitude: Joi.number().float().optional(),
-    latitude: Joi.number().float().optional(),
+    longtitude: Joi.number().optional(),
+    latitude: Joi.number().optional(),
 })
 
 export {
     registerTailorValidation
 }
+
